move multiplex format and wrappers into new multiplex_utils file.
