+++ deepcell_toolbox/deep_watershed.py
@@ -203,97 +203,6 @@ def deep_watershed_mibi(model_output,
     return label_images
 
 
-def deep_watershed_subcellular(model_output, compartment='whole-cell', whole_cell_kwargs=None,
-                               nuclear_kwargs=None):
-    """Postprocess model output to generate predictions for distinct cellular compartments
-
-    Args:
-        model_output (dict): Output from deep watershed model. A dict with a key corresponding to
-            each cellular compartment with a model prediction. Each key maps to a subsequent dict
-            with the following keys entries
-            - inner-distance: Prediction for the inner distance transform.
-            - outer-distance: Prediction for the outer distance transform
-            - fgbg-fg: prediction for the foreground/background transform
-            - pixelwise-interior: Prediction for the interior/border/background transform.
-        compartment: which cellular compartments to generate predictions for.
-            must be one of 'whole_cell', 'nuclear', 'both'
-        whole_cell_kwargs (dict): Optional list of post-processing kwargs for whole-cell prediction
-        nuclear_kwargs (dict): Optional list of post-processing kwargs for nuclear prediction
-
-    Returns:
-        numpy.array: Uniquely labeled mask for each compartment
-
-    Raises:
-        ValueError: for invalid compartment flag
-    """
-
-    valid_compartments = ['whole-cell', 'nuclear', 'both']
-
-    if whole_cell_kwargs is None:
-        whole_cell_kwargs = {}
-
-    if nuclear_kwargs is None:
-        nuclear_kwargs = {}
-
-    if compartment not in valid_compartments:
-        raise ValueError('Invalid compartment supplied: {}. '
-                         'Must be one of {}'.format(compartment, valid_compartments))
-
-    if compartment == 'whole-cell':
-        label_images = deep_watershed_mibi(model_output=model_output['whole-cell'],
-                                           **whole_cell_kwargs)
-    elif compartment == 'nuclear':
-        label_images = deep_watershed_mibi(model_output=model_output['nuclear'],
-                                           **nuclear_kwargs)
-    elif compartment == 'both':
-        label_images_cell = deep_watershed_mibi(model_output=model_output['whole-cell'],
-                                                **whole_cell_kwargs)
-
-        label_images_nucleus = deep_watershed_mibi(model_output=model_output['nuclear'],
-                                                   **nuclear_kwargs)
-
-        label_images = np.concatenate((label_images_cell, label_images_nucleus), axis=-1)
-
-    else:
-        raise ValueError('Invalid compartment supplied: {}. '
-                         'Must be one of {}'.format(compartment, valid_compartments))
-
-    return label_images
-
-
-def format_output_multiplex(output_list):
-    """Takes list of model outputs and formats into a dictionary for better readability
-
-    Args:
-        output_list: list of predictions from semantic heads
-
-    Returns:
-        formatted_dict: dictionary with predictions
-
-    Raises: ValueError if model output list is not len(8)
-    """
-
-    if len(output_list) != 8:
-        raise ValueError('output_list was length {}, expecting length 8'.format(len(output_list)))
-
-    formatted_dict = {
-        'whole-cell': {
-            'inner-distance': output_list[0],
-            'outer-distance': output_list[1],
-            'fgbg-fg': output_list[2][..., :1],
-            'pixelwise-interior': output_list[3][..., 1:2]
-        },
-        'nuclear': {
-            'inner-distance': output_list[4],
-            'outer-distance': output_list[5],
-            'fgbg-fg': output_list[6][..., :1],
-            'pixelwise-interior': output_list[7][..., 1:2]
-        }
-    }
-
-    return formatted_dict
-
-
 def deep_watershed_3D(outputs,
                       min_distance=10,
                       detection_threshold=0.1,
