+++ deepcell_toolbox/deep_watershed_test.py
@@ -104,93 +104,6 @@ def test_deep_watershed_mibi():
             watershed_img = deep_watershed.deep_watershed_mibi(model_output=bad_model_output)
 
 
-def test_deep_watershed_subcellular(mocker):
-    # create dict, with each image having a different constant value
-    base_array = np.ones((1, 20, 20, 1))
-
-    whole_cell_list = [base_array * mult for mult in range(1, 5)]
-    whole_cell_dict = {'inner-distance': whole_cell_list[0],
-                       'outer-distance': whole_cell_list[1],
-                       'fgbg-fg': whole_cell_list[2],
-                       'pixelwise-interior': whole_cell_list[3]}
-
-    nuclear_list = [base_array * mult for mult in range(5, 9)]
-    nuclear_dict = {'inner-distance': nuclear_list[0],
-                    'outer-distance': nuclear_list[1],
-                    'fgbg-fg': nuclear_list[2],
-                    'pixelwise-interior': nuclear_list[3]}
-
-    model_output = {'whole-cell': whole_cell_dict, 'nuclear': nuclear_dict}
-
-    # whole cell predictions only
-    whole_cell = deep_watershed.deep_watershed_subcellular(model_output=model_output,
-                                                           compartment='whole-cell')
-    assert whole_cell.shape == (1, 20, 20, 1)
-
-    # nuclear predictions only
-    nuclear = deep_watershed.deep_watershed_subcellular(model_output=model_output,
-                                                        compartment='nuclear')
-    assert nuclear.shape == (1, 20, 20, 1)
-
-    # both whole-cell and nuclear predictions
-    both = deep_watershed.deep_watershed_subcellular(model_output=model_output,
-                                                     compartment='both')
-    assert both.shape == (1, 20, 20, 2)
-
-    # make sure correct arrays are being passed to helper function
-    def mock_deep_watershed_mibi(model_output):
-        pixelwise_interior_vals = model_output['pixelwise-interior']
-        return pixelwise_interior_vals
-
-    mocker.patch('deepcell_toolbox.deep_watershed.deep_watershed_mibi', mock_deep_watershed_mibi)
-
-    # whole cell predictions only
-    whole_cell_mocked = deep_watershed.deep_watershed_subcellular(model_output=model_output,
-                                                                  compartment='whole-cell')
-
-    assert np.array_equal(whole_cell_mocked, whole_cell_dict['pixelwise-interior'])
-
-    # nuclear predictions only
-    whole_cell_mocked = deep_watershed.deep_watershed_subcellular(model_output=model_output,
-                                                                  compartment='nuclear')
-
-    assert np.array_equal(whole_cell_mocked, nuclear_dict['pixelwise-interior'])
-
-    with pytest.raises(ValueError):
-        whole_cell = deep_watershed.deep_watershed_subcellular(model_output=model_output,
-                                                               compartment='invalid')
-
-
-def test_format_output_multiplex():
-
-    # create output list, each with a different constant value across image
-    base_array = np.ones((1, 20, 20, 1))
-
-    whole_cell_list = [base_array * mult for mult in range(1, 8)]
-    whole_cell_list = [whole_cell_list[0],
-                       whole_cell_list[1],
-                       np.concatenate(whole_cell_list[2:4], axis=-1),
-                       np.concatenate(whole_cell_list[4:7], axis=-1)]
-
-    # cre
-    nuclear_list = [img * 2 for img in whole_cell_list]
-
-    combined_list = whole_cell_list + nuclear_list
-
-    output = deep_watershed.format_output_multiplex(combined_list)
-
-    assert set(output.keys()) == {'whole-cell', 'nuclear'}
-
-    assert np.array_equal(output['whole-cell']['inner-distance'], base_array)
-    assert np.array_equal(output['nuclear']['inner-distance'], base_array * 2)
-
-    assert np.array_equal(output['whole-cell']['pixelwise-interior'], base_array * 6)
-    assert np.array_equal(output['nuclear']['pixelwise-interior'], base_array * 12)
-
-    with pytest.raises(ValueError):
-        output = deep_watershed.format_output_multiplex(combined_list[:7])
-
-
 def test_deep_watershed_3D():
     shape = (5, 10, 21, 21, 1)
     inner_distance = np.random.random(shape) * 100
